add: implementacao de JWT

+++ src/main/java/io/github/athirson010/financialPlanning/controller/SaldoController.java
@@ -10,6 +10,7 @@ import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -48,7 +49,7 @@ public class SaldoController {
     }
 
     @GetMapping("consultar-gastos/{data}")
-    public Double getSaltoMensal(@RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate data) {
+    public BigDecimal getSaltoMensal(@RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate data) {
         return service.buscarSaldoMensal(data);
     }
